bundle: Reorganize functions on bundle.c
Move bundle_list specific functions from bundle.c to bundle_list.c.
Keep in bundle.c only generic bundle functions and create a bundle header
file to export them.
Bundle-remove is still left to avoid rebase conflicts.
