+++ src/swupd.h
@@ -10,6 +10,7 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "bundle.h"
 #include "config_loader.h"
 #include "globals.h"
 #include "lib/archives.h"
@@ -313,22 +314,17 @@ extern int link_or_copy_all(const char *orig, const char *dest);
 extern int remove_files_from_fs(struct list *files);
 extern void print_pattern(const char *pattern, int times);
 extern void prettify_size(long size_in_bytes, char **pretty_size);
+extern int link_or_rename(const char *orig, const char *dest);
 
 /* subscription.c */
 struct list *free_list_file(struct list *item);
 struct list *free_bundle(struct list *item);
 extern void create_and_append_subscription(struct list **subs, const char *component);
 extern char *get_tracking_dir(void);
+extern int add_subscriptions(struct list *bundles, struct list **subs, struct manifest *mom, bool find_all, int recursion);
 
 /* bundle.c */
-extern bool is_installed_bundle(const char *bundle_name);
 extern enum swupd_code remove_bundles(struct list *bundles);
-extern enum swupd_code show_bundle_reqd_by(const char *bundle_name, bool server);
-extern enum swupd_code show_included_bundles(char *bundle_name);
-extern enum swupd_code list_installable_bundles();
-extern int add_subscriptions(struct list *bundles, struct list **subs, struct manifest *mom, bool find_all, int recursion);
-enum swupd_code list_local_bundles();
-extern int link_or_rename(const char *orig, const char *dest);
 
 /* verify.c */
 extern enum swupd_code verify_main(void);
